Updated formatting BinPackParameters
No longer uses this rule in any way. Instead adheres to dont be stupid
about it.

Method
 void myFunction(float a, float b, float c)

if BinPackParameters: true and if parameterlist is to long would become
this:
 void myFunction(
     float a,
     float b,
     float c)

if BinPackParameters: false and if parameterlist is to long would become
this:
 void myFunction(float a, float b
     float c)

Both can be regarded as bad, thus we do not enforce this. Instead group
parameters logically and do not be stupid about it.