Massive graphic imp. graphicmodule have sharedcont
Graphics now contains code for making a floating, rotating box with all
the directX implementation and shader communicatoin. The system for now is
that one uses the loadmesh function to load a mesh from the specified file
and gets a int id returned wich is a reference to the buffer that is
stored in a vector on the graphicmain in graphicmodule. (Just like the
Agile course).
I also made the change in the graphicmodule to take the sharedcontext as a
in parameter in the contructor, dont know if this is the way we wanted it
and it should be reviewd...