Free look camera fixed
A free look camera is now up and running. The camera manager in the game
core keeps an eye on the different cameras and updates them if they should
be updated. You can swap camera from freelook to third person by pressing
space quickly (The thirdperson camera is not yet implemented). To change
the direction of the free look camera press left mouse