Debuglogging is now working half the way.
Still have problems when there are to much messages in a short time and
the main application crashes.

If the main application crashes it does not allow for the logger to finish
the logging.

This can be solved by sharedmemory being owned by the logging process and
thus being certain that all memory entered here can be correctly written.