PhysicsModule: can now get collisionpairs from physics
GetCollisionPairs returns a vector of a custom struct
(vector<CollisionPair>). Each CollisionPair has two ints, which are the
IDs of the two bodies that have collided.

WARNING! These IDs are still interal. Shall be changed to directly
represent the entityID of the game entity (No. This does not alter the
fact that the engine is independent from the game. Trust me)