Implemented the start of potentialfields
The potentialfields work in 2d (x and z) and the update of the
potentialfield is very unoptimized for now. The potentialfield component
now contains a potentialfield actor, similar to how render component
works. This actor might be used for other thingies too and doesnt have to
be attached to a field.