PlayerHandler: Major rework to make it work with controller
Lots of things are altered to make player work as a controller. Gives way
more control to the user, but also makes things more complex. Subject to
change if it turns out that this is a bad idea.
Old code still exists, but within a for-loop which should not be accessed
ever.