Potential fields fixed + some comments
Potentialfields now works! When using the potentialfield the unit tests
all the 8 quads around himself + the one he is standing on to se which
have the higher charge value and tries to move to that quad. One problem
that have been detected is when creating new entities the pointer to the
actor will be pointing to the same thing unless a new actor is created for
each new entity. The same problem exsists with rigid body and meshes