Potential field for static obj and group for dynam
Potential fields should mainly be used for static objects even though
dynamic objects are supported. For dynamic objects like enemies and player
a group should be used instead. A AI group uses a set of actors to decide
where a position should be attracted. The creation place for ai groups is not
perfect...