Fixed #19 no longer crashes when running into goal in debug mode
onTrigger callback no longer assumes that both the trigger and the other
actor are rigid bodies. Now checks if other actor is a controller, and if
so it iterates through all controllers to figure out which.
WARNING! Using controllers as triggers will cause the exact same crash. No
reason to do so to begin with however