EnemySpawner: Fixed bug when removing bullets
Sometimes the removal of bullets called for the removal of the same rigid
body several times. Used to crash, since removing a body that has already
been removed and thus does not exist is impossible.
Now the remove-method on the server prints an error message, then returns
without crashing. Not a permanent fix, but is necessary anyways to ensure
the physics module is safe