Fixed losing data bug using Circlebuffer as IPC
If the circlebuffer on one side was inited long before the other and has a
chance to actually produce something before the other one was inited. Then
when the second one was initialized it would overwrite all the metadata
and thus some data would be lost.

The current solution may not be top notch, as it assumes that a specific
part of the memory will probably never be a specific value during
release...

This also reduced the adjusted size of the buffer, and thus some tests
fails, I had to adjust some values for this. Also added a lot of more
comments and text on exception for debuggin.