Changed Head & Head to use relative metadata
When two filemaps writes and reades to the same memory they would save
their own pointer to the mapped part of the memory. Thus incorrectness
could be caused because they exist in different memoryspaces. I'm not sure
that this is an actual problem, however, this new solution may is a bit
more clear, and now the code works.