Fixed Logger module, added circlebuffer/filemap logic
A surprising amount of fixes on small things, such as destructors,
includes and const keyword. However, I was a bit stupid and did everything
at the same time so it was horribly hard to commit as seperate commits.
Lesson to next timme :s

Added the entire logic structure for writing to a buffer and then
returning.
And a seperate thread that reads from the same buffer and writes that to
shared memory in turn.