Alot of stuff added to enemies
The enemies now have a mesh and smaller collision sphere so they stand
closer to the ground. The potentialfield have a little problem but are
currently working the same way as they did in the previous map where the
field is genereated abit above the ground to avoid collision vs it. This
will probably be fixed with sweeps instead of trigger