PF actors can now change potentialfield
The PF actors have a jump distance which they add to theire position if
they are standing at the edge. This means that if the platform they are
trying to jump to is to far away or too shot and too close he might fail
the jump becuase he overshoots the new platform.