Made the scale in transcomp matter + offsety draw
The entityfactory now uses the scale from the transcomp, if one exists, when
creating a entity. Also modified the player creation in
playerhandlerclient to use a function that enables scale read from
blueprint. Also made a offsety in the rendercomponent to enable units to
be drawn with a offset in y so they stand on the ground even though physx
says otherwise