Aiming up and down fixed
The player can now aim up and down and also fire at the correct direction.
The other player will also see the particles in the right way/right
direction. If a entity wants to be locked in x rotation this can be
specified in the render component. This is useful since we change the
orientation to enable aiming but we dont want to change the way the
character is drawn